backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
celery[redis]==5.4.0
redis==5.2.0
  pydantic==2.9.2
  pydantic-settings==2.5.2
  onnxruntime-gpu==1.19.2
  rembg==2.0.60
  vtracer==0.6.11
  Pillow==11.0.0
  numpy==2.1.1
  pymatting==1.1.12
  realesrgan==0.3.0
  basicsr==1.4.2
  gfpgan==1.3.8
  opencv-python==4.10.0.84